% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lotame_audience_affinity.R
\name{lotame_audience_affinity}
\alias{lotame_audience_affinity}
\title{Lotame Audience Affinity}
\usage{
lotame_audience_affinity(audience_id, region_id = NULL,
  with_counts = "false", min_uniques = 1000, min_index = 0,
  max_index = 1e+07, min_affinity = 0, max_affinity = 100,
  min_client_index = 0, max_client_index = 1e+07, with_prices = "true",
  page_count = 5000, page_num = 1, sort_attr = "index30",
  sort_order = "DESC", paginate_query = FALSE)
}
\description{
This function generates a data frame of the affinities for the requested audience.
}

