% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lotame_audience_estimates.R
\name{lotame_audience_reachestimate}
\alias{lotame_audience_reachestimate}
\title{Lotame Audience Reach Estimate}
\usage{
lotame_audience_reachestimate(estimate_id)
}
\description{
This function calls the Lotame API.
}
