% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lotame_audiences.R, R/lotame_heirarchies.R
\name{lotame_audiences}
\alias{lotame_audiences}
\title{Lotame API Call}
\usage{
lotame_audiences(client_id)

lotame_audiences(client_id)
}
\description{
This function calls the Lotame API.

This function calls the Lotame API.
}

