% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lotame_campaign_daily.R
\name{lotame_campaign_daily}
\alias{lotame_campaign_daily}
\title{Lotame Audiences}
\usage{
lotame_campaign_daily(campaign_id, lookback_days = 30)
}
\description{
This function returns a data frame of audiences for the requested client.
}
