% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lotame_client_interactions.R
\name{lotame_client_interactions}
\alias{lotame_client_interactions}
\title{Lotame Client Interactions}
\usage{
lotame_client_interactions(client_id)
}
\description{
This function returns a data frame of all interactions types for a client.
}
