% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lotame_hierarchies.R
\name{lotame_hierarchies}
\alias{lotame_hierarchies}
\title{Lotame Hierachies}
\usage{
lotame_hierarchies(client_id)
}
\description{
This function returns a data frame of hierarchies for the requested client.
}

