% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lotame_hierarchy_nodes.R
\name{lotame_hierarchies_nodes}
\alias{lotame_hierarchies_nodes}
\title{Lotame Hierachy Search}
\usage{
lotame_hierarchies_nodes(search_query, client_id,
  hierarchy_category = "STANDARD", date_range = "LAST_30_DAYS",
  universe_id = 1, page_size = 25000)
}
\description{
This function returns a data frame of hierarchies for the search term.
}
