% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lotame_hierarchies_nodes.R
\name{lotame_hierarchy}
\alias{lotame_hierarchy}
\title{Lotame Hierachies}
\usage{
lotame_hierarchy(hierarchy_id, client_id, depth = 0, inc_stats = "true",
  universe_id = 1)
}
\description{
This function returns a data frame of hierarchies for the requested client.
}

