% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lotame_service_ticket.R
\name{lotame_service_ticket}
\alias{lotame_service_ticket}
\title{Lotame Service Ticket}
\usage{
lotame_service_ticket(call)
}
\description{
This function generates a service ticket for the Lotame API using a path.
}

