% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lotame_audience_stats.R
\name{lotame_audience_stats}
\alias{lotame_audience_stats}
\title{Lotame Audience Affinity}
\usage{
lotame_audience_stats(client_id, client_as_group = TRUE,
  group = "audienceId")
}
\description{
This function calls the Lotame API.
}
