% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lotame_audiences.R
\name{lotame_audiences}
\alias{lotame_audiences}
\title{Lotame Audiences}
\usage{
lotame_audiences(client_id)
}
\description{
This function returns a data frame of audiences for the requested client.
}

